Stop resetting maintainer_can_modify PR input field
We were originally resetting this on every pull request update to avoid
a bug in GitHub that reset it to false whenever the PR was modified.

Recently, this was flagged as an validation error, causing the request to
fail with a 422 Unprocessable Entity HTTP error.

Hopefully the bug this was a work around for is not fixed.